requirements.txt
youtube-transcript-api
google-generativeai
pillow
docx2txt